refactor(storage): extract PendingToolCallStore sub-mixin (SQLite + ABC)
Move the twelve pending-tool-call methods (plus the PTC-owned
_insert_pending_tool_call_unlocked) out of the _agent_run.py residual into a
dedicated PendingToolCallStore bucket: SQLitePendingToolCallStoreMixin and the
backend-neutral PendingToolCallStoreABC. Bodies are byte-identical verbatim
moves.

The four cross-bucket transactions (cancel/expire/update_resolved/
mark_not_applicable) still call the AgentRunStore _..._unlocked run-cascade
privates, now reached via MRO from the co-composed SQLiteAgentRunStoreMixin —
each transaction is moved WHOLE (lock/BEGIN IMMEDIATE + all writes + commit/
rollback) and the privates stay SINGLE, never duplicated.

The composite AgentRunMixin ABC now extends
(AgentRunStoreABC, PendingToolCallStoreABC); storage_base/__init__.py and the
BaseStorage base list are unchanged. SQLitePendingToolCallStoreMixin is composed
ahead of the shrinking residual SQLiteAgentRunMixin.
