+++ reflexio/server/services/storage/sqlite_storage/__init__.py
@@ -24,7 +24,7 @@ from ._stall_state import (
     mark_stall_notified,
     upsert_stall_state,
 )
-from .agent_run import SQLiteAgentRunStoreMixin
+from .agent_run import SQLiteAgentRunStoreMixin, SQLitePendingToolCallStoreMixin
 from .base import SQLiteDeletionMixin, SQLiteFtsVecMixin
 from .governance import (
     AuditEventStoreMixin,
@@ -45,6 +45,7 @@ from .profiles import InteractionStoreMixin, ProfileSearchMixin, ProfileStoreMix
 
 class SQLiteStorage(
     SQLiteAgentRunStoreMixin,
+    SQLitePendingToolCallStoreMixin,
     SQLiteAgentRunMixin,
     ProfileStoreMixin,
     InteractionStoreMixin,
