+++ reflexio/server/services/storage/sqlite_storage/agent_run/__init__.py
@@ -1,3 +1,4 @@
 from ._agent_run_store import SQLiteAgentRunStoreMixin
+from ._pending_tool_call_store import SQLitePendingToolCallStoreMixin
 
-__all__ = ["SQLiteAgentRunStoreMixin"]
+__all__ = ["SQLiteAgentRunStoreMixin", "SQLitePendingToolCallStoreMixin"]
