+++ reflexio/server/services/storage/storage_base/agent_run/__init__.py
@@ -1,3 +1,4 @@
 from ._agent_run_store import AgentRunStoreABC
+from ._pending_tool_call_store import PendingToolCallStoreABC
 
-__all__ = ["AgentRunStoreABC"]
+__all__ = ["AgentRunStoreABC", "PendingToolCallStoreABC"]
